Added command line options to specify interface (-i <addr>) and IPv6 interface index (-x <index>) Updated to use getaddrinfo for name lookup - using the supplementary function included into common library when using MUSICA IPv6. Re-enabled the fprintf to print warnings u as dialogue boxes under Windows.

+++ video/grabber-win32.cpp
@@ -801,7 +801,7 @@ void VfwGrabber::start()
 
 	/*1e6 / double(max_fps_);*/
 	/*(DWORD)frametime_*/
-	parms_.dwRequestMicroSecPerFrame = 1e6 / double(max_fps_);
+	parms_.dwRequestMicroSecPerFrame = (unsigned long) (1e6 / double(max_fps_));
 	parms_.wPercentDropForError = 100;
 	parms_.fUsingDOSMemory = FALSE;
 	parms_.wNumVideoRequested = 3;
